CSS style work
Added grid to header component. Also added a grid function to big btns with names and icons and then a sub grid for a sub group of btns with just icons. Working on importing svg files into WSL to add to my project. After that I will be making the header more responsive and will be adding animations to the header section.