Refactored nav-links
BgLink and SmLink are now the same element. CSS values are changed based on a bool value that is passed into the component. Need to go back and get higher res pngs for fallback function. Fallback function now works. onError a png will load incase an svg dose not work on an older browser