Wrote some unit test for nav-links.jsx
Wrote a unit test to make sure the svg loads when no error is present and the pngs load an img onError. Working on a test to make sure the svgs and pngs exist in the project folder and are in the correct dir. Then I will work on adding a test to ensure that the nav-links take you to the right page url.