Working on adding graphs and error function
Working on adding resuable responsive graphs via chart.js. Also working on a function that throws an error if the value entered is undfinded. This is to be used in componets that require certain params to work properly. A error saying what parama is missing will make debugging a lot easier.