Added to generateTrend
Working on getting genrateTrend working. Is meant to return an array of dates with currency values in a x y format. This is done so it can be readable to line chart. For some reason the front of the array is return undefined for x and NAN for y. The end of the array has the dates and currency ammount. Working on fixing this bug