Got genreateTrend working
genrateTrend now has x and y keys so that line chart can read it. The values are not undefined anymore. The problem was that I was deleating the old keys before the new ones where made. Might need to make a copy of the old key val and asign it so that i can deleate the old key val in order to keep the object clean.