Worked on some test and did some research on DOM trees
Did some research on if it is better to remove an item from a dom or just do display none. Found that for componets with out large sub DOM trees it is better to remove them from the DOM and just rerender them later. I am attempting to create a wrapper componet to handle these toggles in and out of the DOM. I am also working on making this wrapper compatiple with animations so if I want I can add dropdown animations or fade in/out animations to the children of this wrapper. Then after this I will finish writing my test for the dashboard layout. I initially started working on this part of my code because I wanted my component to not be in the DOM at a certain point while testing and I relized that display none dose not acctually remove the compoent from the DOM.